Generalised and moved the find_nearest_point function from perturbation_auf_fortran to misc_function_fortran, as it will also need to be used in photometric_likelihood. Various test re-calls and setup.py changes to propagate the change.
