+++ macauff/tests/test_misc_functions.py
@@ -11,48 +11,15 @@ import scipy.special
 from ..misc_functions import (create_fourier_offsets_grid, load_small_ref_ind_fourier_grid,
                               hav_dist_constant_lat)
 from ..misc_functions_fortran import misc_functions_fortran as mff
-from .test_misc_functions_fortran import haversine_wrapper
+from .test_shared_library_fortran import haversine_wrapper
 
 
-def test_haversine_formula():
-    a, b, c, d = 0, 0, 0, 0
-    e = haversine_wrapper(a, b, c, d)
-    assert e == 0
-
-    a, b, c, d = 0.0, 1/3600, 0, 0
-    e = haversine_wrapper(a, b, c, d)
-    assert_allclose(e, 1/3600)
-
-    a, b, c, d = 0.0, 1/3600, 2.45734, 2.45734
-    e = haversine_wrapper(a, b, c, d)
-    assert_allclose(e, 1/3600*np.cos(np.radians(c)))
-
-    a, b, c, d = 25.10573, 25.30573, 15.48349, 15.48349
-    e = haversine_wrapper(b, a, c, d)
-    assert_allclose(e, 0.2*np.cos(np.radians(c)))
-
-    a, b, c, d = 51.45734, 51.45734, 0, 1/3600
-    e = haversine_wrapper(a, b, c, d)
-    assert_allclose(e, 1/3600)
-    e = haversine_wrapper(a, b, d, c)
-    assert_allclose(e, 1/3600)
-
-    c, d = 86.47239, 85.47239
-    e = haversine_wrapper(a, b, d, c)
-    assert_allclose(e, 1)
-
-    a, b, c, d = 0, 180, -45, 45
-    e = haversine_wrapper(a, b, c, d)
-    assert_allclose(e, 180)
-
-    a, b, c, d = 10, 250, -50, 65
-    e = haversine_wrapper(a, b, c, d)
-    sin_half_lat = np.sin(np.radians((c - d)/2))
-    cos_lat_1 = np.cos(np.radians(c))
-    cos_lat_2 = np.cos(np.radians(d))
-    sin_half_lon = np.sin(np.radians((a - b)/2))
-    assert_allclose(e, 2 * np.degrees(np.arcsin(np.sqrt(sin_half_lat**2 +
-                    cos_lat_1*cos_lat_2*sin_half_lon**2))))
+def test_closest_auf_point():
+    test_points = np.array([[1, 1], [50, 50]])
+    source_points = np.array([[3, 2], [80, 50], [40, 30]])
+    inds = mff.find_nearest_point(source_points[:, 0], source_points[:, 1],
+                                  test_points[:, 0], test_points[:, 1])
+    assert np.all(inds == np.array([0, 1, 1]))
 
 
 def test_calc_j0():
