+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import find_packages
 from numpy.distutils.core import Extension, setup
 
-names = ['perturbation_auf_fortran', 'group_sources_fortran', 'misc_functions_fortran']
+names = ['group_sources_fortran', 'misc_functions_fortran']
 f90_args = ["-Wall", "-Wextra", "-Werror", "-pedantic", "-fbacktrace", "-O0", "-g", "-fcheck=all",
             "-fopenmp"]
 
@@ -10,8 +10,8 @@ extension = [Extension(name='macauff.{}'.format(name), sources=['macauff/{}.f90'
                        extra_f90_compile_args=f90_args, libraries=['shared_library'])
              for name in names]
 
-extension.extend([Extension(name='macauff.tests.test_misc_functions_fortran',
-                            sources=['macauff/tests/test_misc_functions_fortran.f90'],
+extension.extend([Extension(name='macauff.tests.test_shared_library_fortran',
+                            sources=['macauff/tests/test_shared_library_fortran.f90'],
                             language='f90', extra_link_args=["-lgomp"],
                             libraries=['shared_library'], extra_f90_compile_args=f90_args)])
